--- a/package.json
+++ b/package.json
@@ -24,8 +24,8 @@
     "test": "cross-env NODE_OPTIONS=--max-old-space-size=8192 pnpm run ut",
     "test:all": "pnpm run test && pnpm run e2e",
     "sort-package-json": "npx sort-package-json \"packages/*/package.json\"",
-    "ut": "vitest run --pool=forks",
-    "ut:watch": "vitest"
+    "ut": "rstest run",
+    "ut:watch": "rstest"
   },
   "nano-staged": {
     "*.{md,mdx,json,css,less,scss}": "prettier --write",
@@ -40,6 +40,7 @@
     "@changesets/cli": "^2.29.5",
     "@modern-js/module-tools": "^2.67.11",
     "@rsdoctor/tsconfig": "workspace:*",
+    "@rstest/core": "0.0.9",
     "@scripts/test-helper": "workspace:*",
     "cross-env": "^7.0.3",
     "cspell-ban-words": "^0.0.4",
@@ -47,9 +48,7 @@
     "husky": "^9.1.7",
     "nano-staged": "^0.8.0",
     "nx": "^20.8.2",
-    "prettier": "^3.5.3",
-    "vite-tsconfig-paths": "4.3.2",
-    "vitest": "3.0.9"
+    "prettier": "^3.5.3"
   },
   "pnpm": {
     "overrides": {