--- a/packages/ai/tests/code.test.ts
+++ b/packages/ai/tests/code.test.ts
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import path from 'path';
-import { expect, test } from 'vitest';
+import { expect, test } from '@rstest/core';
 import { ChatGPT } from './model/openai';
 import { codePrompt } from '@/prompt/code';