--- a/packages/ai/tests/socket.test.ts
+++ b/packages/ai/tests/socket.test.ts
@@ -1,4 +1,4 @@
-import { expect, test } from 'vitest';
+import { expect, test } from '@rstest/core';
 import { createSocket, sendRequest } from '../src/server/socket';
 import { SDK } from '@rsdoctor/types';