--- a/packages/ai/tests/tsconfig.json
+++ b/packages/ai/tests/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "../tsconfig.json",
-  "include": [".", "../vitest.setup.ts"],
+  "include": [".", "../rstest.setup.ts"],
   "exclude": ["fixtures/**/*"]
 }