--- a/packages/components/package.json
+++ b/packages/components/package.json
@@ -17,7 +17,7 @@
     "build": "modern build",
     "dev": "npm run start",
     "start": "modern build -w",
-    "test": "vitest run"
+    "test": "rstest run"
   },
   "exports": {
     ".": {