--- a/packages/core/tests/build/utils/loader/basic.test.ts
+++ b/packages/core/tests/build/utils/loader/basic.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest';
+import { describe, it, expect } from '@rstest/core';
 import { extractLoaderName } from '@/build-utils/build/utils';
 
 describe('test src/build/utils/loader.ts basic functions', () => {