--- a/packages/core/tests/build/utils/loader/changeBuiltinLoader.test.ts
+++ b/packages/core/tests/build/utils/loader/changeBuiltinLoader.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest';
+import { describe, it, expect } from '@rstest/core';
 import os from 'os';
 import { Plugin } from '@rsdoctor/types';
 import { Utils } from '@/build-utils/build';