--- a/packages/core/tests/build/utils/loader/loadLoaderModule.test.ts
+++ b/packages/core/tests/build/utils/loader/loadLoaderModule.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest';
+import { describe, it, expect } from '@rstest/core';
 import { resolve } from 'path';
 import { loadLoaderModule } from '@/build-utils/build/utils';