--- a/packages/core/tests/build/utils/loader/mapEachRules.test.ts
+++ b/packages/core/tests/build/utils/loader/mapEachRules.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest';
+import { describe, it, expect } from '@rstest/core';
 import { mapEachRules } from '@/build-utils/build/utils';
 
 describe('test src/build/utils/loader.ts mapEachRules()', () => {