--- a/packages/core/tests/plugins/config.test.ts
+++ b/packages/core/tests/plugins/config.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest';
+import { describe, it, expect } from '@rstest/core';
 import { RuleSetRule } from 'webpack';
 import { makeRulesSerializable } from '@/inner-plugins/utils'