--- a/packages/graph/tests/utils.ts
+++ b/packages/graph/tests/utils.ts
@@ -1,4 +1,4 @@
-import { beforeAll, afterAll, afterEach, beforeEach } from 'vitest';
+import { beforeAll, afterAll, afterEach, beforeEach } from '@rstest/core';
 import { File, Server } from '@rsdoctor/utils/build';
 import { Common, SDK } from '@rsdoctor/types';
 import { request } from 'http';