--- a/packages/sdk/tests/server/apis/index.test.ts
+++ b/packages/sdk/tests/server/apis/index.test.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it } from 'vitest';
+import { describe, expect, it } from '@rstest/core';
 import { SDK } from '@rsdoctor/types';
 import { Router } from '../../../src/sdk/server/router';