--- a/packages/sdk/tests/server/utils.test.ts
+++ b/packages/sdk/tests/server/utils.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest';
+import { describe, it, expect } from '@rstest/core';
 import { getDataByPagination } from '../../src/sdk/server/utils';
 
 describe('test server/utils.ts', () => {