--- a/packages/utils/tests/common/algorithm.test.ts
+++ b/packages/utils/tests/common/algorithm.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest';
+import { describe, it, expect } from '@rstest/core';
 import { Algorithm } from '../../src/common';
 
 describe('test src/common/algorithm.ts', () => {