--- a/packages/utils/tests/common/bundle.test.ts
+++ b/packages/utils/tests/common/bundle.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest';
+import { describe, it, expect } from '@rstest/core';
 import { Bundle } from '../../src/common';
 
 describe('test src/common/bundle.ts', () => {