--- a/packages/utils/tests/common/graph.test.ts
+++ b/packages/utils/tests/common/graph.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest';
+import { describe, it, expect } from '@rstest/core';
 import { Graph } from '../../src/common';
 
 describe('test src/common/graph.ts', () => {