--- a/packages/utils/tests/common/url.test.ts
+++ b/packages/utils/tests/common/url.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest';
+import { describe, it, expect } from '@rstest/core';
 import { Url } from '../../src/common';
 
 describe('test src/common/url.ts', () => {