--- a/packages/utils/tests/json.test.ts
+++ b/packages/utils/tests/json.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest';
+import { describe, it, expect } from '@rstest/core';
 import { Json } from '../src/build';
 
 describe('test src/json.ts', () => {