--- a/packages/utils/tests/parser/parser.test.ts
+++ b/packages/utils/tests/parser/parser.test.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it } from 'vitest';
+import { describe, expect, it } from '@rstest/core';
 import { parser } from '../../src/rule-utils/parser';
 
 describe('test src/rule-utils/parser/parser.ts', () => {