--- a/packages/webpack-plugin/package.json
+++ b/packages/webpack-plugin/package.json
@@ -16,7 +16,7 @@
     "dev": "npm run start",
     "build": "modern build",
     "start": "modern build -w",
-    "test": "vitest run"
+    "test": "rstest run"
   },
   "dependencies": {
     "@rsdoctor/core": "workspace:*",