--- a/packages/webpack-plugin/tests/multiple.test.ts
+++ b/packages/webpack-plugin/tests/multiple.test.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it } from 'vitest';
+import { describe, expect, it } from '@rstest/core';
 
 /**
  * create sandbox to load src/multiple.ts to avoid sdk save in global variable between different test cases.