--- a/scripts/test-helper/package.json
+++ b/scripts/test-helper/package.json
@@ -2,7 +2,7 @@
   "name": "@scripts/test-helper",
   "version": "0.1.1",
   "private": true,
-  "description": "The vitest helpers of Rsdoctor.",
+  "description": "The rstest helpers of Rsdoctor.",
   "repository": {
     "type": "git",
     "url": "https://github.com/web-infra-dev/rsdoctor",