… packages/ai/vitest.config.ts
… packages/core/tests/build/utils/loader/__snapshots__/changeBuiltinLoader.test.ts.snap
… packages/core/tests/build/utils/loader/__snapshots__/changeBuiltinLoaderVue.test.ts.snap
… packages/core/tests/common/module-graph/__snapshots__/transform.test.ts.snap
… packages/core/tests/plugins/__snapshots__/circle-check.test.ts.snap
… packages/core/tests/plugins/__snapshots__/config.test.ts.snap
… packages/graph/tests/__snapshots__/module-graph.test.ts.snap
… packages/graph/tests/__snapshots__/utils.test.ts.snap
… packages/sdk/tests/__snapshots__/module-graph.test.ts.snap
… packages/utils/tests/__snapshots__/crypto.test.ts.snap
… packages/utils/tests/__snapshots__/json.test.ts.snap
… pnpm-lock.yaml
… vitest.workspace.ts