## Summary

Use Rstest as Rsdoctor's test framework. 🌟
- use rstest instead of vitest.
- use `source.tsconfigPath` instead of `vite-tsconfig-paths` dependency. ( 👈  core performance improvement point)
- use `source.decorators` instead of tsconfig `experimentalDecorators` option.

before:
```bash
npx vitest run  20.13s user 6.76s system 326% cpu 8.244 total
```

after:
```bash
npx rstest run  15.77s user 4.83s system 368% cpu 5.595 total
```

<img width="1874" height="322" alt="image" src="https://github.com/user-attachments/assets/f44610ef-b93f-4df6-97e3-6fa7c8e31365" />

<img width="1652" height="248" alt="image" src="https://github.com/user-attachments/assets/71e7009c-b142-42ed-9408-cff1df205b35" />


## Related Links

https://rstest.rs/

https://rsbuild.rs/guide/migration/vite#vite-tsconfig-paths
<!--- Provide links of related issues or pages -->